Update wca-regulations-2012.md
A3d3: I must have missed that change before. Why would we change this? "12 seconds" is long and unfamiliar. It is better to have a short signale at this point of the preinspection.
A6h: The penalty made no sense there. With that wording it would be a DNF if ANYONE made a move or an alignment. If a competitor does so, then this case is covered by A6e.