Explictly allow medical aids to be electronic. Addresses #231.
In this case, 2i2a) lists an exception to 2i), which is two levels higher (the "grandparent" Regulation). All other "inline" exceptions apply only to the current or the parent Regulation. In the case of medical aids, I think the exception is sufficiently fair and clear, and unlikely to cause confusion.