"Systematic leniency": allow waiving penalties under certain conditions. Addresses #798.
Ideally, we would like to learn when Delegates believe that it is more fair to waive certain penalties instead of being more strict. This gives them a mechanism to do so. The change requires noting this in the Delegate report so that we can understand which Regulations may need clarifications or revisions in order to ensure fair rulings across the world.