Clearing up the use of sight block in Blindfolded Solving
We're normally taught that the should only be used after the competitor applies the first move to the puzzle. This is in order to follow B4e, where a competitor may undon the blindfold before they've made the first move.

However, B4c2 states the jugde should place the objetc while the competitor is wearing the blindfold. If we follow this, after the competitor has donned the blindfold, but before he has made his first move, the judge should still be using the blocker.

This is unnecessary, as the competitor can go back to the memorization phase and see the puzzle any time he wants, and clearly doesn't reflect what happens in competitions.

Besides, B4c1 states that the competitor must wear the blindfold such that their view of the puzzle would still clearly be blocked if the sight blocker were not in the way. Therefore using a sight blocker before the competitor has applied the first move to the puzzle is unnecessary.

This commit removes B4c2 and the proposed guideline, and instead adapts the wording of the Regulation B4c.

In this way, we make it clearer in the Regulations when the sight blocker should be used, without adding a Guideline when it is not necessary.