Improving the wording of Regulation 11g
As pointed out in issue #874, Regulation 11g currently doesn't make it clear that Delegates have to ensure copies of the Regulations and Guidelines are available to all competitors and officials, and not only to himself.

Rather than adding a Guideline to clear this up, we can fix this by simply adding the Regulations must be available to "officials and compeitors" with the appropriate changes in propositions and wording to make this work.