Update Regulation A6e to WRC understanding
Update Regulation A6e to WRC understanding by specifying the requirements for lifting the time penalty for touching the puzzle after releasing it.

Fixes #200, #534, [this](https://forum.worldcubeassociation.org/t/regulation-a6e-revisement/3225/3) Forum post and makes the adjustment mentioned on Incident Log [#23](https://www.worldcubeassociation.org/incidents/23).

There are three main issues that should be addressed on this update:
1. When can the competitor touch the puzzle after stopping the timer - the current wording "until the judge has inspected the puzzle" is not very clear;
2. On what circumstances can the time penalty for touching the puzzle after stopping the timer be lifted; and
3. Whether a puzzle falling on the competitor's lap should still be penalized - currently made clear by Guideline A6e+.

The solution for each one of these issues is:

1. Making it so the competitor can only touch the puzzle after the attempt ends (after the competitor has signed the scoresheet, as per Regulation A7c) or after being explicitly allowed to do so by the judge.

2. Instituting the following three requirements for lifting the time penalty for touching the puzzle after releasing it (also present in the above mentioned Incident Log):
- The touch must happen immediately after stopping the timer - this Regulation aims at being lenient toward instinctual or automatic reactions (like the one made on [this solve](https://www.youtube.com/watch?v=YSo8w1yPetE&feature=youtu.be&t=1h22m12s) by Lucas Etter), and not touch to the puzzle that happens when the competitor is calm or can think things through;
- The touch must be brief - for the same reason as above, a prolonged touch to the puzzle constitutes a clear violation of Regulation A6e; and
- The touch must not affect the state of the puzzle - if it does, the attempt should be disqualified, as per the proposed Regulation A6e1. To be able to determine whether this happened or not, the WCA Delegate may use video evidence, as per Regulation 11f.

This commit also adds the explicit need for WCA Delegate discretion for applying the exception, as judges shouldn't be the ones making these decisions.

I believe an update do A6e shouldn't get into the merit of the touch being intentional or not. Besides it being subjective to determine intention (you could argue, for example, that Lucas' touch in the above-linked video was intentional), it isn't the best criteria - sometimes an unintentional touch should be penalized (e.g. a puzzle falling on a competitor's lap) and an "intentional" (or instinctual) touch shouldn't.

3. As a puzzle falling on a competitor's lap does not configure a brief touch - thus not meeting the requirements defined in Regulation A6e2 - it should still be penalized. That's why Guideline A6e+ should still stand as it is.

Finally, there shouldn't be any discretion in determining whether a violation of Regulation A6e results in a DNF or a +2. That's why these two situations were treated separately and made different.