Clarify how to handle results of disqualifiied attmpts
This change clarifies how should the results of disqualified attempts be handled.

It comes as a result of [incident log #58](https://www.worldcubeassociation.org/incidents/58), in which a competitor was disqualified at the start of a round before having finished any attempts. In alignment with Regulation 2j1a, the Delegate recorded all of his attempts as DNFs.

However, this artificially increases the number of competitors in a round (see Regulation 9m1-9m3) and misrepresents what happened in practice.

The proposed wording for 2j1a makes it clear that the rest of the remaining attempts should only be recorded as DNFs if the competitor has finished one or more attempts. Otherwise, if the competitor has finished no attempts, as is the case in the aforementioned incident log, no results should be recorded.