Improve wording on the results of disqualified attempts
This commit improves the wording on the handling of results of attempts when a competitor is disqualified from a specific event.

The wording was changed from "finished" to "started" so that if a competitor is disqualified while in the middle of their first attempt, their results get recorded as DNF, in alignment with the definition of Regulation 9f4 of what constitutes a DNF.

I chose "started at least one attempt" so that we don't get any weird wording from "started one or more attempts in a round".

I changed "after being disqualified" to "before being disqualified", as it is what was originally intended and was incorrectly committed before.

Finally, I changed "has" to "had" because I believe it is a more appropriate verb tense in the context.