Clarify who can approve competitors not participating in all events of a competition
This commit clarifies who can approve competitors not participating in all events of a competition.

As highlighted by @Epride [here](https://forum.worldcubeassociation.org/t/inaccuracy-of-regulation-9k/8926),  Regulation 9k currently inaccurately states that the WCA Board is the one responsible for approving that competitors aren't able to participate in all events of a competition.

However, in almost all cases, this approval is not carried out by the Board, but by the WCAT.

This commit, therefore, makes it clear that the Board specifically doesn't need to give the mentioned approval.

We chose not to change the wording in the Regulation mentioning specifically the WCAT so that the Regulations doesn't need to be updated if the Team changes its name or if anoother entity in the WCA becomes responsible for the approval.