Add guideline B4c+ (BLD sight blocker timing) (#952)
* add guideline B4c+ (BLD sight blocker timing)

The judge should put up the sight blocker after the competitors have applied the first move, because competitors can remove the blindfold if they have not applied any moves.

* Clearing up the use of sight block in Blindfolded Solving

We're normally taught that the should only be used after the competitor applies the first move to the puzzle. This is in order to follow B4e, where a competitor may undon the blindfold before they've made the first move.

However, B4c2 states the jugde should place the objetc while the competitor is wearing the blindfold. If we follow this, after the competitor has donned the blindfold, but before he has made his first move, the judge should still be using the blocker.

This is unnecessary, as the competitor can go back to the memorization phase and see the puzzle any time he wants, and clearly doesn't reflect what happens in competitions.

Besides, B4c1 states that the competitor must wear the blindfold such that their view of the puzzle would still clearly be blocked if the sight blocker were not in the way. Therefore using a sight blocker before the competitor has applied the first move to the puzzle is unnecessary.

This commit removes B4c2 and the proposed guideline, and instead adapts the wording of the Regulation B4c.

In this way, we make it clearer in the Regulations when the sight blocker should be used, without adding a Guideline when it is not necessary.

* Update wca-regulations.md

Co-authored-by: Lucas Garron <lgarron@worldcubeassociation.org>

Co-authored-by: Pedro Miranda <pmiranda@worldcubeassociation.org>
Co-authored-by: Iván Brigidano Pérez <47502355+Ivan-Brigidano@users.noreply.github.com>
Co-authored-by: Lucas Garron <lgarron@worldcubeassociation.org>