Improve the wording of Regulation A6e after feedback
This commit improves the proposed wording of Regulations A6e, A6e1, and A6e2 after WCA Staff and Community Feedback.

Summary of changes:
1. Change back the moment until the competitor is forbidden from touching the puzzle to when the "judge has inspected the puzzle", as it is currently
2. Improve the wording of "moving" to "applying a move"
3. Change WCA Delegate discretion to judge discretion in regards to who can lift the penalty of touching the puzzle
4. Change the wording of "may not" to "may" in Regulation A6e2

On a general note, the WRC appreciated the proposal that consists of applying no penalty if the judge was able to determine the state of the puzzle before the competitor touched it and applying a DNF if the judge wasn't able to do it.

However, we decided against this proposal as it lessens the punishment/discouragement for competitors that touch the puzzle after finishing their attempt. In our view, this would lead to more and more competitors "greying" the line of when they can and cannot touch the puzzle, which is something we definitely want to steer away from.

Therefore, we agreed on maintaining a +2 in these Regulations, encouraging competitors to not touch the puzzle until the judge has been able to clearly inspect the puzzle.

With that being said, here are the details on the specific changes of this commit:

1. After WCA Staff feedback, it was clear that requiring competitors not to touch the puzzle until they had signed the scoresheet was unnecessarily strict and had no good reason to be this way. I had attempted to solve two problems at once but ended up going in an undesired direction.

2. This improvement comes from @SAuroux very accurately pointing out that "moving" a puzzle is not the same as "applying a move" to it. The wording change, therefore, is meant to make the proposed wording clearer.

3. This change is a result of @KitClement correctly highlighting that, in most cases, it is not the WCA Delegate that will be deciding to lift the +2 penalty, but the judge.

4. This modification was made after @xsrvmy indicated that "may not" is not up to RFC 2119 standard.

Thank you to everyone who contributed to this update!