Clarify how to handle results of disqualified attempts (#996)
* Clarify how to handle results of disqualifiied attmpts

This change clarifies how should the results of disqualified attempts be handled.

It comes as a result of [incident log #58](https://www.worldcubeassociation.org/incidents/58), in which a competitor was disqualified at the start of a round before having finished any attempts. In alignment with Regulation 2j1a, the Delegate recorded all of his attempts as DNFs.

However, this artificially increases the number of competitors in a round (see Regulation 9m1-9m3) and misrepresents what happened in practice.

The proposed wording for 2j1a makes it clear that the rest of the remaining attempts should only be recorded as DNFs if the competitor has finished one or more attempts. Otherwise, if the competitor has finished no attempts, as is the case in the aforementioned incident log, no results should be recorded.

* Improve wording on the results of disqualified attempts

This commit improves the wording on the handling of results of attempts when a competitor is disqualified from a specific event.

The wording was changed from "finished" to "started" so that if a competitor is disqualified while in the middle of their first attempt, their results get recorded as DNF, in alignment with the definition of Regulation 9f4 of what constitutes a DNF.

I chose "started at least one attempt" so that we don't get any weird wording from "started one or more attempts in a round".

I changed "after being disqualified" to "before being disqualified", as it is what was originally intended and was incorrectly committed before.

Finally, I changed "has" to "had" because I believe it is a more appropriate verb tense in the context.

* Improve wording of Regulation 2j1a

As per @Ivan-Brigidano 's request, this commit improves the wording of Regulation 2j1a so that it is consistent with the wording of the previous sentence and the rest of the Regulations.

* Update wca-regulations.md

Co-authored-by: Iván Brigidano Pérez <47502355+Ivan-Brigidano@users.noreply.github.com>