Update Regulation A6e to WRC understanding (#983)
* Update Regulation A6e to WRC understanding

Update Regulation A6e to WRC understanding by specifying the requirements for lifting the time penalty for touching the puzzle after releasing it.

Fixes #200, #534, [this](https://forum.worldcubeassociation.org/t/regulation-a6e-revisement/3225/3) Forum post and makes the adjustment mentioned on Incident Log [#23](https://www.worldcubeassociation.org/incidents/23).

There are three main issues that should be addressed on this update:
1. When can the competitor touch the puzzle after stopping the timer - the current wording "until the judge has inspected the puzzle" is not very clear;
2. On what circumstances can the time penalty for touching the puzzle after stopping the timer be lifted; and
3. Whether a puzzle falling on the competitor's lap should still be penalized - currently made clear by Guideline A6e+.

The solution for each one of these issues is:

1. Making it so the competitor can only touch the puzzle after the attempt ends (after the competitor has signed the scoresheet, as per Regulation A7c) or after being explicitly allowed to do so by the judge.

2. Instituting the following three requirements for lifting the time penalty for touching the puzzle after releasing it (also present in the above mentioned Incident Log):
- The touch must happen immediately after stopping the timer - this Regulation aims at being lenient toward instinctual or automatic reactions (like the one made on [this solve](https://www.youtube.com/watch?v=YSo8w1yPetE&feature=youtu.be&t=1h22m12s) by Lucas Etter), and not touch to the puzzle that happens when the competitor is calm or can think things through;
- The touch must be brief - for the same reason as above, a prolonged touch to the puzzle constitutes a clear violation of Regulation A6e; and
- The touch must not affect the state of the puzzle - if it does, the attempt should be disqualified, as per the proposed Regulation A6e1. To be able to determine whether this happened or not, the WCA Delegate may use video evidence, as per Regulation 11f.

This commit also adds the explicit need for WCA Delegate discretion for applying the exception, as judges shouldn't be the ones making these decisions.

I believe an update do A6e shouldn't get into the merit of the touch being intentional or not. Besides it being subjective to determine intention (you could argue, for example, that Lucas' touch in the above-linked video was intentional), it isn't the best criteria - sometimes an unintentional touch should be penalized (e.g. a puzzle falling on a competitor's lap) and an "intentional" (or instinctual) touch shouldn't.

3. As a puzzle falling on a competitor's lap does not configure a brief touch - thus not meeting the requirements defined in Regulation A6e2 - it should still be penalized. That's why Guideline A6e+ should still stand as it is.

Finally, there shouldn't be any discretion in determining whether a violation of Regulation A6e results in a DNF or a +2. That's why these two situations were treated separately and made different.

* Fix mistakes in initial PR

Fixes trailling whitespace and missing link in a reference to Regulation 11f.

* Improve the wording of Regulation A6e after feedback

This commit improves the proposed wording of Regulations A6e, A6e1, and A6e2 after WCA Staff and Community Feedback.

Summary of changes:
1. Change back the moment until the competitor is forbidden from touching the puzzle to when the "judge has inspected the puzzle", as it is currently
2. Improve the wording of "moving" to "applying a move"
3. Change WCA Delegate discretion to judge discretion in regards to who can lift the penalty of touching the puzzle
4. Change the wording of "may not" to "may" in Regulation A6e2

On a general note, the WRC appreciated the proposal that consists of applying no penalty if the judge was able to determine the state of the puzzle before the competitor touched it and applying a DNF if the judge wasn't able to do it.

However, we decided against this proposal as it lessens the punishment/discouragement for competitors that touch the puzzle after finishing their attempt. In our view, this would lead to more and more competitors "greying" the line of when they can and cannot touch the puzzle, which is something we definitely want to steer away from.

Therefore, we agreed on maintaining a +2 in these Regulations, encouraging competitors to not touch the puzzle until the judge has been able to clearly inspect the puzzle.

With that being said, here are the details on the specific changes of this commit:

1. After WCA Staff feedback, it was clear that requiring competitors not to touch the puzzle until they had signed the scoresheet was unnecessarily strict and had no good reason to be this way. I had attempted to solve two problems at once but ended up going in an undesired direction.

2. This improvement comes from @SAuroux very accurately pointing out that "moving" a puzzle is not the same as "applying a move" to it. The wording change, therefore, is meant to make the proposed wording clearer.

3. This change is a result of @KitClement correctly highlighting that, in most cases, it is not the WCA Delegate that will be deciding to lift the +2 penalty, but the judge.

4. This modification was made after @xsrvmy indicated that "may not" is not up to RFC 2119 standard.

Thank you to everyone who contributed to this update!

Co-authored-by: Iván Brigidano Pérez <47502355+Ivan-Brigidano@users.noreply.github.com>