Relax the blindfolded logo restriction
A logo on a center piece only impacts blindfolded solving if the logo can be detected while blindfolded, which means it is either engraved, a sticker, or otherwise changing the texture of that center piece. If the logo is smoothly part of the plastic with no detectable texture/elevation changes, then it does not provide additional information while blindfolded and should be allowed.