9f1: Remove "nearest" from wording (#1107)
Rule 9f1 currently says that all timed results under 10 minutes are
truncated to the "nearest" hundredth of a second. This is confusing as
when you truncate, you don't apply any concepts of "nearest". Both X.000
and X.009 would be recorded as X.00 if one were to truncate the
precision at the hundredths place. The nearest hundredth to X.009 is
X.01.
This resolves #1106

---------

Co-authored-by: Lucas Garron <code@garron.net>