Make 5b5 more consistent (#1120)
## Relevant Information

**This proposal is an expansion of Viktor Zenk's idea.**

5b5f+ includes a key definition. Please, read it before proceeding.

There are (at least) two possible courses, one of them should be added
explicitly to the wording:

1. **Valid positions are absolute**: The idea is that valid positions
are not relative to the state of other pieces. Therefore, It does not
matter if the nearest valid position is currently "claimed" by other
piece, it is still valid when considering the final position of the
affected piece.
2. **Valid positions are relative**: Valid positions are determined
depending on the current state of other pieces.

### Example
In the following example, the possible paths described above have
different outcomes:
1. DNF. The nearest valid positions for the affected corner are the
positions currently occupied by the blue-red-yellow and
orange-blue-yellow corners. Such positions do not render the piece
solved.
2. OK. As other positions are occupied, the only valid position for the
corner is the solved one (it is the only position available where a
corner may be normally placed).

![screen_shot_2023-04-13_at_5 49
23_pm_720](https://user-images.githubusercontent.com/59543882/231909036-3b7124db-dd07-4955-ab82-f866e6991dfd.png)

---------

Co-authored-by: Carter Kucala <42009313+CarterKoala@users.noreply.github.com>
Co-authored-by: Antonio Kam <19kamh3@gmail.com>