Hotfix for rounded average in 9f2
9f2 was amended for the 2025-01-01 Regulations release to replace rounding with truncation. The motivation behind that change was driven by individual results. Justification and documentation for the change focused on individual results, and did not discuss means and averages.

The updated wording changed 9f2 from rounding to truncation for both single results and means and averages. Following software implementation, the change for means and averages was considered undesirable, as it results in averages and means being calculated differently depending on when they were achieved.

Since the undesired change was not fully intended, and did not form part of the justification presented to Staff and the Community, an out-of-cycle hotfix is considered appropriate.